Move the higher-order vector commands to a new rust module
