+++ src-tauri/src/command/mod.rs
@@ -1,4 +1,5 @@
 
+pub mod accum;
 pub mod algebra;
 pub mod arguments;
 mod base;
@@ -199,8 +200,8 @@ pub fn default_dispatch_table() -> CommandDispatchTable {
   map.insert("vmask".to_string(), Box::new(BinaryFunctionCommand::named("vmask")));
   map.insert("norm".to_string(), Box::new(vector::NormCommand::new()));
   map.insert("cross".to_string(), Box::new(BinaryFunctionCommand::named("cross")));
-  map.insert("vapply".to_string(), Box::new(vector::VectorApplyCommand::new()));
-  map.insert("vmap".to_string(), Box::new(vector::VectorMapCommand::new()));
+  map.insert("vapply".to_string(), Box::new(accum::VectorApplyCommand::new()));
+  map.insert("vmap".to_string(), Box::new(accum::VectorMapCommand::new()));
 
   // Vector statistics commands
   map.insert("mean".to_string(), Box::new(dispatch_on_flags_command(FlagDispatchArgs {
